plan: Added Bazel support for the Plan interface language
Plan is a minimal Lisp designed to be used to specify the ABI of
an operating system. With the ABI defined in a Plan document,
compatible code can be generated to either call or answer the
corresponding system calls. This ensures that user and kernel
code do not have different interpretations of the same data in
memory.

This adds basic Bazel support for Plan documents, which can be
checked, formatting checked, and used to generate user or kernel
code in Rust.

See bazelbuild/rules_rust#485 and bazelbuild/rules_rust#689.

For now, generated Rust files cannot be included as normal source
files in a rust_library, as they exist in the wrong directory so
rustc does not find them. The work-around for now is to use an
environment variable and the include!(env!(ENV_VAR)) combination.
It's not pretty, but it works. However, this then breaks rust_doc
targets, as they do not inherit the rust_library's data field, so
are unaware of the dependency, which Bazel (rightly) rejects.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>