plan: Ensure the type system is aware of the syscalls enumeration
We generate a synthetic enumeration for the complete set of system
calls, but we only do so once a Plan document has been fully type
checked. This means that a document could define a type with the
same name (which would be overwritten) and that any references to
the synthetic enumeration would be treated as a non-existant type.

This commit fixes that by including the synthetic enumeration in
the set of type definitions before parsing the Plan document. This
will both prevent the creation of a duplicate type and will ensure
any references to the enumeration succeed correctly.

To ensure this behaviour continues, this also adds a test for each
property.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>