plan: Generate syscall_handler and type-check syscalls in kernel
This is a small but important improvement to Plan. Previously, we
would generate the set of enumerations and structures in a Plan
document (including the synthetic enumeration for the set of syscalls)
for the kernel, but not the set of syscalls themselves. This meant
that the syscall implementations could misinterpret each syscall's
arguments or results. This is due to the fact that it is much easier
to generate code to call syscalls with the right signature than it
is to ensure a function being called has the right signature.

This commit fixes that. Where previously the syscall handler (and
syscalls) was implemented by hand, it is now generated. The syscall
implementations themselves are now included as associated functions
in a generated trait. The generated syscall handler then calls each
function as appropriate. Using a trait ensures the syscall
implementations are properly type checked and using associated
functions should ensure that there is no unnecessary overhead when
calling the functions, as there is no trait object to box.

As with the generated kernel code previously, this is somewhat more
coupled to the surrounding code than would be ideal. The test case
(which is now more detailed and has full coverage) and the Rust
compiler's type checking should ensure that the generated code is
correct and safe.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>