plan: Generate HTML documentation from a Plan document
This commit adds (and uses) functionality to generate HTML documentation
from a Plan document. This produces an index page summarising the
document and individual pages for each enumeration, structure, and
syscall. Each reference to an enumeration or structure links to its
definition. Pages render in either light or dark mode, depending on
the device preference. Currently, there is no way to override this.

The colour scheme is simple; blue for enumerations, green for structures,
and red for syscalls. Types are printed in their lower case, spaced form.
For now, the individual HTML files are generated and stored in a folder.

The only technical detail of note is that the CSS for each component is
stored in a separate template and included where needed in each HTML
template. This maximises performance unnecessarily.

Future work will generate a ZIP archive containing the full documentation
bundle, add a sidebar navigation to the subpages, and add richer formatting
to type documentation.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>