plan: Swap types.Docs from a slice of strings to slice of DocsItem
We want to support more complex formatting in documentation with
native support in the Plan language, rather than parsing the contents
of the raw strings. This commit takes the next step down that path by
switching types.Docs from []string to []DocsItem, with Text being the
only implementation of DocsItem. This puts in place the foundations
for other types down the line without changing the current functionality.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>