plan: Handle documentation references to syscalls
As Plan does not see syscalls as being a data structure, they do
not satisfy the Type interface. This prevents references to syscalls
in documentation using the `reference` formatting expression.

This commit fixes that issue by creating the SyscallReference type,
which only contains the syscall's name. The type checking code and
tests have been updated accordingly.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>