plan: Ensure the kernel code uses the right error type and order
This fixes two issues. Firstly, the kernel places any generic error
(such as an illegal enumeration parameter) in the first return value
if the syscall only has one result. Secondly, when returning generic
errors, it uses the correct error type for the syscall in question.

This means that a syscall that has an error enumeration with different
numerical values for the generic error values will return the correct
result.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>