plan: Always pass the error enumeration in result2
This is an ABI change. Previously, a syscall that defined only one
result would pass that result in the first result value (register
RAX). This complicated some error handling, as generic errors the
syscall handler returned were passed in either RAX or RDX, depending
on the individual syscall.

This commit fixes that by passing the error enumeration in the
second result unconditionally, returning 0 in the first result.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>