plan: Replace synthetic enumeration for integers with Docs method
Rather than having a static data structure describing the integer
types, we can make the functionality dynamic by providing a Docs
method on the Integer type, which returns the documentation for
the specified integer type. This means that the user does not need
to find the type they need in the values.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>