plan: Check the syscall number earlier in the syscall handler
In the future, we will want to be able to perform checks on the
syscall being invoked, before we prepare to invoke it. To simplify
these checks, we now convert to the Syscalls enum (or handle a
non-existant syscall) straight away, then match on the enum.

This does not yet make any functional changes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>