plan: Error enumerations must start with required values
This makes the requirements on error enumerations slightly more
strict. Rather than requiring an error enumeration to contain the
set of required values (in any order), it must now start with them,
in order. This simplifies error handling when a syscall is rejected
with `bad_syscall`, as it's guaranteed to have the same numerical
value, even if the syscall is not defined and therefore its error
type is undefined.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>