vendeps: Added dependency vendoring tooling
It's all very well using tools to pull in third-party libraries,
such as rules_rust's crate_repositories and rules_go's go_repository,
they can be painful to use (building cargo-bazel requires pulling
in a lot of crates and takes the majority of the build time on a
fresh build) and make it easy to ignore the weight of the dependency
graph. They can also make it harder to analyse and understand the
dependency graph.

vendeps is tooling to manage, download, and build our third-party
library dependencies. It uses a dependency specification written
in a Starlark file, then uses it to download the dependencies to
the vendor folder, using cache information to avoid downloading
dependencies we already have. The results are stored into a
dependency manifest, which is stored in vendor/manifest.bzl. The
vendoring process also uses the information in the specification
to generate BUILD files for each dependency. Note that unlike
the tools it replaces, it generates test targets for the dependencies
it vendors.

One of the biggest real-world changes this will cause is that when
we change the dependency graph, the difference in code will be
reflected in the Git diff. This makes it easier to see the impact
of a change and gives us some chance to spot malicious or otherwise
unwanted changes.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>