vendeps: Added functionality to update the dependency specification
Now that we have removed the functionality for updating third-party
library dependencies from //tools/update-deps, we need a replacement
option. This commit solves that problem by adding the functionality
to //tools/vendeps, as that already has code to parse and understand
the dependency specification in deps.bzl.

The approach is very similar to the previous functionality removed
from //tools/update-deps. The dependency specification in deps.bzl
is parsed, we use crates.io and proxy.golang.org to identify any
dependencies that can be updated, and perform the updates by making
changes to the deps.bzl AST and the writing the updated structure
back out to disk. Note that this does not then vendor the changed
dependencies, in case the user wants to make manual changes first.

The main difference when compared with the old functionality from
//tools/update-deps is that when a major update to a stable Rust
crate is available, it is logged but ignored. This is to reduce
the chance of an incompatible change being attempted. For the
purposes of the definition above, a major change is where the
version changes its major version (the v1 in v1.2.3) and a stable
crate is one where the major version is not v0. Note that we do
not perform the same check for Go modules, as Go already treats
major updates as separate modules, so the same effect occurs
implicitly.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>