cpu: Enable Supervisor Mode Execution Protection if available
This enables SMEP, which prevents the execution of usermode code
(code mapped in with the U/S flag set) when not running as ring 3.
The kernel does not currently run usermode code, so this should
not affect any existing behaviour.

For more details on SMEP, see x86_64 manual volume 3A, section
4.6.1.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>