thread: Set the Access Control RFLAG when switching to user mode
To enable Supervisor Mode Access Prevention (SMAP), we need to
ensure that when switching to user mode (and therefore using a
user stack while still executing in ring 0), we set the Access
Control (AC) flag in the RFLAGS register. This ensures that if
SMAP is enabled, we won't trigger a page fault.

This takes place in three places. Firstly, we set the AC flag
just before switching stacks when we're about to switch to a user
stack. This is because although our first action after switching
stack is to load the RFLAGS register, we do so by popping its
value off the stack, which results in a read from user memory.
Secondly, we set the AC flag in the saved RFLAGS register when
we create a user thread. This is so that when we pop the RFLAGS
register off the stack as described above, we maintain the AC
flag and can continue to access the user stack. Finally, we unset
the AC flag once we have restored the kernel stack when we receive
a syscall. This ensures that if SMAP is enabled, the syscall
handler will be affected by SMAP as expected.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>