syscalls: Opt in to user memory access in syscall implementations
To enable Supervisor Mode Access Prevention (SMAP), we must ensure
that any deliberate access to user memory is flagged as such, to
prevent SMAP from triggering a page fault. As a result, when we
deliberately access user memory, we use the new with_user_memory_access
functionality to set the Access Control flag in the RFLAGS register
and opt in to temporary access to user memory.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>