cpu: Enable Supervisor Mode Access Prevention if available
This enables SMAP when supported by the CPU. This means that any
unintended access to user memory when running in ring 0 will result
in a page fault. To opt in to user memory access, the Access Control
(AC) flag in the RFLAGS register can be set. This is already done
when switching to a user thread and in the syscall implementations
that access user memory.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>