smap: Moved SMAP enabling for syscalls from assembly to Rust code
We re-enable SMAP when returning from userspace to the kernel in
a syscall so that the syscall handler and implementations are fully
affected by SMAP. However, doing this in assembly limits the options
available to us. Although it means slightly more code runs without
SMAP protection, this gives us more flexibility.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>