plan: Added the ability to generate a shared Rust module
Whereas the kernel and user modules generate code specific to each
use case, this generates just the shared code that is the same for
both.

Future commits will remove the generation of the same code in the
user and kernel code, instead importing the data structures from
the shared module.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>