plan: Separated syscall implementations/callers from data structures
Although this touches a lot of files, it's not a huge change.

Previously, the ABI was split cleanly into two parts; the syscall
implementations in the kernel and the helper functions for calling
them in userspace. In both cases, the data structures and types
were also defined. This commit follows on from 7f0a302b in splitting
them apart.

Now, the data structures and types are defined for both use cases
in //shared/firefly_abi. This code is also an entire crate where
previously it was (part of) a module. As a result, we can now generate
Rust documentation for it, as we need not worry about generated and
non-generated Rust files being in different directories, as we only
have generated code now.

The crate above is then imported by the kernel and user code, rather
than its contents being generated afresh in both. This makes the kernel
and user code easier to read, as it only has the code unique to that
use case.

In the process of making this change, a few other changes have been
made which makes the code easier to understand, use, and maintain.

Firstly, the user code is in now its own crate in //user/firefly_syscalls.
As with //shared/firefly_abi, this makes some code management easier
and we can now generate Rust docs for it. This also makes some naming
less ambiguous.

Secondly, the generated code for the kernel has been renamed to
firefly_syscalls to avoid clashing terminology, but it remains a
module within //kernel. This is necessary, as there is a cyclical
dependency with the kernel itself, so the syscalls cannot be their
own crate.

Thirdly, with all the data structures going into the shared library,
the templates within //tools/plan/encoding/rust are now more simple,
which also meant moving some test code around.

One important nuance to moving the shared data structures into their
own crate in //shared/firefly_abi is that they can now be imported
and used throughout the kernel. Previously, as they were only defined
(for the kernel) in the root kernel crate, they couldn't be used in
any subcrates within the kernel. This hadn't caused any problems yet,
but it would have done eventually.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>