plan: Always specify all 6 args in userspace syscall helper functions
Previously, we would only specify whichever args were directly used
by each syscall. This was an (unnecessary) optimisation that meant
that a syscall that took only (say) 3 arguments would leave the
remaining 3 registers with unspecified values.

Unnecessary optimisation aside, this results in a brittle design,
as the kernel cannot enforce any constraints on the unused argument
registers. To allow such an improvement, this commit ensures that
unused arguments are always provided with the value 0.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>