plan: Enforce that unused syscall args must have value 0
Following on from 07ff2e97, this allows the syscall ABI to be
more strict, which will help reduce compatibility issues. If a
given argument is not used by a syscall and its value is not 0,
the kernel will now treat that as an illegal value and return
an error accordingly. This means that buggy user code cannot
supply junk values and be caught out if an extra argument starts
to be used.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>