plan: Always return a Result in user helper functions for syscalls
Previously, the helper functions we generate to allow user processes
to call syscalls would return a Result if the syscall returned two
results and a bare Error (or variant thereof) if the syscall returned
only one result. This was unnecessarily inconsistent, so we now return
Result<(), Error> instead.

This resulted in some changes to the user code, but nothing else has
changed.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>