plan: Added support for new integer types
This adds support for the 'new type' idiom, where a unique type
is defined, with an integer underlying type. This can be used to
ensure that common integral types, such as file handles or port
numbers can be defined once and type-checked thereafter, including
ensuring that the type has a consistent size.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>