bootloader: Added copyright headers, clarified licensing
This clarifies that all work in the original bootloader crate was
copyright 2018 Philipp Oppermann (whose work is much appreciated)
and that subsequent changes are copyright 2022 (or later) The
Firefly Authors.

The original work remains available under the MIT license, but
the modified codebase is licensed under the same BSD 3-clause
license as the rest of the project.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>