virtmem: Only parse the kernel binary when necessary
When we're remapping the kernel, we only need to identify the
page mappings we don't need any more, which are those in the
null pages, userspace, and the boot info. We don't need the same
level of granularity that we do when debugging, so there's no
need to parse the kernel binary so we can break down the page
mappings within the binary in more detail.

We still go to the full extent when debugging the page mappings.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>