bootloader: Separate the kernel from the bootloader
This is a fairly big change.

Previously, the kernel was linked into the bootloader when the
bootloader was compiled, using a Cargo build script. This allowed
the linker script to insert symbols for the kernel's start address
within the bootloader binary and the size of the kernel in size.
This worked effectively, but it meant that the bootloader had to
be recompiled every time the kernel changed. The process of linking
the kernel into the bootloader also made the build quite fragile,
as it relied on implicit details like the location of the modified
kernel binary produced by the build script.

This changes the architecture somewhat. Now, the bootloader's linker
script produces an address where the kernel should exist (as before,
but now written explicitly in the linker script, rather than added
in a build script) and an address within the stage 2 bootloader
where the kernel's size will exist at runtime. The bootloader has
been adjusted to read the kernel size by dereferencing the symbol,
rather than the symbol containing the size.

This means that the bootloader can now be compiled as a standalone
binary, without needing access to the kernel, and without needing a
build script. This will make future changes simpler and means we should
not need to recompile the bootloader unless it (or its dependencies)
change.

To enable this change, the bootimage tool is now a little more
advanced. It takes both the bootloader and the kernel binary and
concatenates them into the bootable image. The last little detail
is that it writes the size of the kernel into the bootloader binary
within the bootable image. This means that the kernel can change
size and the bootloader will see the updated size at runtime, without
needing to recompile the bootloader.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>