bootinfo: Separate out bootinfo from bootloader
The kernel shouldn't need to depend on the bootloader, particularly
when the bulk of the bootloader is in a separate crate (the binary
crate). This commit recognises that by moving the bootinfo details
into a separate, shared crate. This will allow the bootloader to be
rewritten to move the bulk of the functionality into the library
crate, resulting in a similar structure to the kernel and making it
easier to use clippy and other lints.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>