bootloader: Refactor binary crate into library crate, restructure
This is a non-functional change. Now that nothing outside //bootloader
depends on the library crate, we can move the vast majority of the
bootloader's functionality into the library crate, resulting in a
structure much more similar to //kernel. The main difference is that
the entry point is not defined in Rust (_start is in stage 1), so
we cannot define it in main.rs.

Having moved all this functionality, we can now add rust_doc and
and other extra build targets, which use the library crate as their
connection point. We also move the x86_64_bare_metal_rust_binary
target from the workspace root into //bootloader.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>