vendeps: Remove unused Go packages from used modules
Whereas Rust has only one library crate per package, a Go module
can contain many packages. Currently, we just ignore the extra
code, as it isn't used. However, it does waste space in the repo,
particularly golang.org/x/tools, which contains a substantial
amount of code, only a tiny amount of which is used here.

After we have downloaded and unzipped each module, we now scan
the resulting file tree for directories that do not exist in the
packages declared in the dependency specification and delete them.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>