thread: Moved initial stack population into separate functions
When a new kernel or user thread is created, its initial stack
contents are populated. Note that this is not true for an idle
thread, as it inherits its stack.

Populating the stack is quite a subtle operation, with lots of
opportunities for edge cases and off-by-n errors. This commit
moves the code into separate methods on StackBounds, making it
much easier to test. This also adds some initial tests.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>