vendeps: Treat unstable versions as less stable
On reflection, a change in unstable version is more likely to
constitute a 'major update' than a change in stable version.
To reflect that, this commit makes two key changes. Firstly,
a v0.x to v0.y update is now major if x > y. This reflects
how semver is supposed to work. Secondly, a change from v0
to v1 is also treated like a major update, just like any other
change of major version.

This is less likely to result in unwanted updates and we need
not worry too much about false negatives, as the first such
version is still logged.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>