plan: Extracted common per-item HTML into separate templates
Previously, we had a lot of duplicated HTML structure at the
beginning and end of each item template, such as the HTML head
tag and the title. This has now been moved into two new templates,
meaning that more of each item template is unique to that template.

This will make it easier to insert multiple items into a single
page in a consistent manner.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>