plan: Added support for groups of items to aid documentation
It can sometimes be difficult to identify related bits of
functionality in the ABI. This commit adds a new kind of
declaration that groups together such related functionality.
A group consists of the usual name and documentation, plus
one or more items (new integers, enumerations, bitfields,
structures, or syscalls). When documentation is generated,
the group will bring together the details for all its items
in one place.

Note that this has no effect on generated code, which has
no need for such a concept.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>