memory: Added address ranges for user binary, heap, stack regions
By splitting userspace into three chunks, it becomes easier to
ensure that the data from one region does not accidentally spill
over into another. As the heap is the region likely to be the
largest, it gets the larger share. The nearly 128 TiB userspace
region is therefore split into:

- user binary (less than 32 TiB)
- user heap (64 TiB)
- user stack (32 TiB)

All three regions are expected to be sparse and to be populated
in randomised chunks (except for static executables), thanks to
ASLR.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>