plan: Simplified checking whether a type fits in a register
We need to know whether a given type fits in a register so that
we can detect and prevent a Plan that produces an invalid ABI
that tries to pass an invalid type through a register in a
syscall.

Previously, we did this with a type switch, listing the set of
types that fit in a register. This functioned fine, but it was
too easy to define a new type and forget to add it to the list.

The simpler approach is for the Type interface to include a new
method, Register, which returns whether the Type is allowed in
a register. While the current set of types do not vary their
answer based on the architecture, that may change in the future,
so we provide the current architecture to the method.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>