plan: Make type titles in documentation more flexible
We use the toItemUnderlyingType helper function to declare when
a type should be documented as having an underlying type. Previously,
this worked by returning the underlying type (or nil for types with
no underlying type, such as a structure). This was then printed in
the item-prefix template using an if statement.

This works fine, but only if the underlying type is a simple type.
To allow more complex types, this commit switches to toItemUnderlyingType
returning the full HTML for the extra information, or an empty string
for types with no underlying type. This means we can be more flexible
in future.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>