plan: Sort types code alphabetically
This is a non-functional change. Previously, code handling Plan
types was in an arbitrary order of integers, enumerations,
bitfields, arrays, structures, syscalls, then groups. This commit
simply reorders that to put the itme types in alphabetical order.
That is, the order becomes arrays, bitfields, enumerations, integers,
structures, syscalls, groups. Groups are kept last, as they are
metadata, rather than being part of the ABI.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>