plan: Added integration tests for generated code
Currently, we have tests to check that the code generated from a
Plan document is the code we expect. However, a change in the
semantics of that language could still result in a user process
and the kernel using a different representation for the same
data, or misinterpreting the output from the other.

This commit adds integration tests to detect that event, in the
unlikely scenario that it happens. This is done by generating
code for a sample Plan document and inserting tests in that
language that confirm that the generated data structures have
the expected layout in memory. We then use Bazel rules to
compile and test the generated code.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>