plan: Move Rust gentests into separate folder, use standard Plan
This commit simplifies the code involved in the gentests. Previously,
we used a custom Go binary to compile a Plan document and append
hand-written tests to the generated code. This worked fine, but
it was more complex than necessary. This commit undoes some of
that complexity.

Now, instead of using a custom binary to generate the code and
append the tests to the generated code, we instead use standard
Plan Bazel rules to generate Rust code.

We the build a separate crate to run all the tests. This means
we can keep the two Rust codebases separate and don't need to
accommodate one another's idiosyncrasies.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>