abi: Merged ABI definition and shared generated code
This is a comparatively minor structural change that sets us up
for the future much better. Previously, we had the Plan document
for the Firefly ABI in //shared/syscalls and the shared Rust
code generated in //shared/firefly_abi. This worked fine but it
seemed odd to have the ABI definition in a different place from
"firefly_abi". Secondly, it worked nicely for the Rust code we
use today but would have been clumsy to add support for other
languages in the future.

This solves those problems by moving the Plan document (and its
Bazel targets) into //shared/firefly_abi and by renaming the
Rust targets for the generated code to "rust*". This means that
we can add generated code for other languages unambiguously.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>