tar: Added the ability to start reading at a block offset
While we will normally want to read TAR archives from the start
of a block device, we will sometimes want to read part-way into
a device. This adds a simple version of such functionality by
adding a second constructor that takes an offset into the
sequence of blocks in the device. The resulting TAR reader will
ignore any blocks before this offset. This means that the
original constructor (`new`) is behaves identically to the new
constructor (`new_at_offset`) with an offset of 0.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>