image: Include userspace into the disk image
Previously, we would construct one disk image for the MBR record,
bootloader, and kernel, and another for the userspace (TAR archive).
Now, we merge the two into a single disk image. This will make
deployments simpler, as we only need the single disk image.

So that the kernel can identify the userspace in the root disk
image, we write the offset into the disk image where the userspace
image begins to the last 4 bytes before the 2-byte MBR magic at
the end of the first sector.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>