virtio: Switch to simpler MMIO accesses for PCI transport
Previously, we used the read_volatile! and write_volatile!
macros to access offsets into the common configuration data
of the PCI configuration space.

This commit switches to a slightly simpler approach, which
uses the read and write methods on the MMIO region, using
constant offsets produced by memoffset::offset_of!.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>