ruse: Add the set of Ruse language tokens
This is the start of the Ruse language, which will be a simple
language for writing low-level software without an operating
system. The intent is that this will be used to replace Rust
in the Firefly kernel.

Ruse uses a simple syntax inspired by LISP. The most notable
features of the syntax include:

- Annotations, which can be prepended to any list expression.
  Annotations are a general-purpose mechanism for adding
  structured information, such as architecture requirements,
  metadata, and ABI information. An annotation is a quoted
  list, where the first element is always an identifier.
- No support for floating point numbers. Ruse is designed
  for low-level programming, which should not require the
  use of floating point arithmetic.
- Many structural elements borrowed from Go, such as mandatory
  package identifiers, similar package functionality, case-
  based API visibility, and similar file handling.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>