ruse: Add a fork of golang.org/x/arch/x86/x86spec
This adds an updated and extended fork of the
golang.org/x/arch/x86/x86spec tool. The tool takes
an Intel x86 manual in PDF form and parses the
document. The resulting information is used to
produce structured information about the x86
instruction set. The results are written out in
a CSV file.

The majority of changes in the fork consist of
minor tweaks to support newer versions of the PDF
than was previously used. However, some log messages
are now fatal errors, to aid debugging. Finally,
the tool now records any tuple type of each instruction.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>