ruse: Use the generated x86 data to add instructions to x86
This adds another build target for //tools/ruse/internal/x86,
which incorporates the structured data containing the x86
instruction set. This will be very useful later.

Note that structured instruction data uses the UID field,
which is deliberately chosen to be a valid identifier in most
programming languages and to be unique to a single instruction
form.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>