ruse: Add a tool to generate x86 instruction test data
Writing an assembler requires good test data, particularly
for a complex instruction set like x86. Unfortunately, as
most production-ready x86 assemblers were written a long
time ago (and in C, where unit tests are not widely used),
there does not seem to be much prior art for x86 assembly
test cases.

To solve this, this commit adds a tool to produce such
test data. This works by enumerating the x86 instructions
in //tools/ruse/internal/x86, procedurally generating x86
syntax, resulting in a large set of assembly examples
(currently just over 2.5 million cases). Then, we use
Clang to assemble the tests and Objdump to disassemble
the results, so that we can be confident that the result
is correct and match up the assembly to the machine code.

We then use various heuristics to confirm that this is
correct (as both tools have various quirks). Finally,
the results are written out in CSV format.

Note that this tool is not expected to be run often, as
it requires external dependencies not managed by Bazel
(clang and objdump). Instead, it can be run when needed
and the result checked in. This commit does not yet
check in that data, but it will follow.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>