ruse: Add a partial Ruse compiler and assembler
This adds a few combined parts. The first is a compiler
to transform a Ruse package in abstract syntax tree and
type information form into SSAFIR. The second is an x86
assembler, using structured information about the instruction
set from //tools/ruse/internal/x86.

It is worth noting testdata/x86-test.csv, which has been
generated with //tools/ruse/internal/x86/gen-tests. This
consists of over 2.5 million test cases, mapping Intel
and Ruse x86 assembly and a CPU mode to machine code in
hexadecimal form, along with the UID of the instruction
form.

The assembler tests then use this data, iterating through
the test cases, assembling the Ruse assembly form and
checking that the resulting machine code matches the
machine code in the test case. Currently, those tests all
pass!

The compiler is fairly basic, in that it does not support
function calls, or many other features, but it does
basically work. The assembler is much more robust. It
implements just over 5 thousand x86 instruction forms,
for 16-, 32-, and 64-bit CPU modes. It also includes a
feature to use an annotation to force a particular
instruction form, even if another form would be more
compact. This feature would have been very useful for
generating the tests, had Clang had a similar feature.

The main feature missing from the assembler is support
for jump labels and calls to named functions.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>