ruse: Generate x86 instruction data explicitly, not as dependencies
Currently, we execute the following sequence of tools, managed by
Bazel as dependencies:

- http_file (to fetch the Intel x86 manual, volume 2)
- x86csv (to parse the manual and produce x86 instruction data as a CSV)
- x86json (to transform and enrich x86.csv and output x86.json and data.go)

This works fairly well, but it means that the generated instruction
data is hidden by Bazel and not visible in Git. This may make it harder
to debug changes when the manual is updated. For example, if an instruction's
formatting in the manual changes, and it is missed by x86csv, we would lose
its implementation, perhaps without realising.

This commit makes two main changes. Firstly, x86json becomes an explicit
tool, invoked using `bazel run`, rather than being invoked in the background
in a `genrule`. Secondly, we now write data.go and x86.json to the version-tracked
repository, rather than Bazel's cache. This doesn't change anything downstream,
but it makes both files visible to Git.