ruse: Add type-checking for some special forms in assembly functions
We want to have support in assembly functions for a limited set of
constant special forms, such as `len`, `string-pointer`, and `func`.
This will allow us to reference other parts of a Ruse programme from
assembly. For example, it would allow us to call a Ruse (or assembly)
function from assembly with the following syntax:

```
	; ...
	(call (func other-function))
	; ...
```

This commit adds type checking for those forms.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>